Move PublishOptions to parent package
